requirements.txt
python-telegram-bot==20.7
mercadopago==2.2.2
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx~=0.25.2
